start version 3.1.0-SNAPSHOT

+++ security/org.beangle.security.blueprint/src/main/java/org/beangle/security/blueprint/function/service/FuncPermissionService.java
@@ -55,7 +55,7 @@ public interface FuncPermissionService {
    * 
    * @param roleId
    */
-  Set<String> getResourceNamesByRole(Long roleId);
+  Set<String> getResourceNamesByRole(Integer roleId);
 
   /**
    * 按照角色查找资源
